+++ shacl_testsuite/src/manifest.rs
@@ -1,10 +1,21 @@
 use std::collections::HashSet;
 use std::path::Path;
+<<<<<<< HEAD
 
 use iri_s::IriS;
 use oxrdf::Term;
 use shacl_validation::runner::GraphValidatorRunner;
 use shacl_validation::runner::ValidatorRunner;
+=======
+use std::str::FromStr;
+
+use indoc::formatdoc;
+use oxigraph::io::RdfFormat;
+use oxigraph::model::{GraphNameRef, NamedNode};
+use oxigraph::{model::Term, store::Store};
+use oxiri::Iri;
+use shacl_ast::ShaclParser;
+>>>>>>> 597911f3ae39befd00a88821b84ec56cdb56faf6
 use shacl_validation::shacl_validation_vocab;
 use shacl_validation::validation_report::report::ValidationReport;
 use srdf::RDFFormat;
@@ -22,7 +33,43 @@ pub trait Manifest<S: SRDF + SRDFBasic> {
     where
         Self: Sized;
 
+<<<<<<< HEAD
     fn load_data_graph(path: &Path, base: &str) -> S;
+=======
+impl Manifest {
+    fn new(base: String, store: Store, includes: Vec<Manifest>, entries: HashSet<Term>) -> Self {
+        Manifest {
+            base,
+            store,
+            includes,
+            entries,
+        }
+    }
+
+    // TODO: Change load_graph by load_from_read
+    #[allow(deprecated)]
+    pub fn collect_tests(&self) -> Result<Vec<ShaclTest>, ManifestError> {
+        let mut ans = Vec::new();
+        for entry in &self.entries {
+            let query = formatdoc! {
+                "
+                    SELECT ?action ?result ?label
+                    WHERE {{
+                        {} {} ?action .
+                        {} {} ?result .
+                        OPTIONAL {{ {} {} ?label }}
+                    }}
+                ",
+                entry, shacl_validation_vocab::MF_ACTION.as_named_node(), // check it is blank node
+                entry, shacl_validation_vocab::MF_RESULT.as_named_node(), // check it is not literal
+                entry, srdf::RDFS_LABEL.as_named_node(),
+            };
+
+            let solution = match select(&self.store, query) {
+                Ok(solution) => solution,
+                Err(_) => break,
+            };
+>>>>>>> 597911f3ae39befd00a88821b84ec56cdb56faf6
 
     fn base(&self) -> String;
 
@@ -68,6 +115,7 @@ pub trait Manifest<S: SRDF + SRDFBasic> {
             )?
             .unwrap();
 
+<<<<<<< HEAD
             let shapes_graph_iri = get_object_for(
                 self.store(),
                 &action,
@@ -77,6 +125,25 @@ pub trait Manifest<S: SRDF + SRDFBasic> {
 
             let shapes = Self::format_path(shapes_graph_iri.to_string());
             let data = Self::format_path(data_graph_iri.to_string());
+=======
+            // TODO: explain this
+            let term = data_graph_iri.to_string();
+            let mut chars = term.chars();
+            chars.next();
+            chars.next_back();
+            let path = chars.as_str().to_string();
+
+            let mut data_store = self.store.clone(); // explicit copy
+            if path != self.base {
+                data_store = Store::new()?;
+                data_store.bulk_loader().load_graph(
+                    BufReader::new(File::open(path.replace("file:/", ""))?),
+                    RdfFormat::Turtle,
+                    GraphNameRef::DefaultGraph,
+                    Some(&self.base),
+                )?;
+            }
+>>>>>>> 597911f3ae39befd00a88821b84ec56cdb56faf6
 
             ans.push(ShaclTest::new(
                 data,
@@ -93,17 +160,49 @@ pub trait Manifest<S: SRDF + SRDFBasic> {
         Ok(ans)
     }
 
+<<<<<<< HEAD
     fn load(path: &Path) -> Result<Self, ManifestError>
     where
         Self: Sized,
     {
         let base = match Path::new(path).canonicalize()?.to_str() {
+=======
+    // TODO: Change load_graph by load_from_read
+    #[allow(deprecated)]
+    pub fn load(file: &str) -> Result<Manifest, ManifestError> {
+        let path = Path::new(file);
+
+        let base = match path.canonicalize()?.to_str() {
+>>>>>>> 597911f3ae39befd00a88821b84ec56cdb56faf6
             Some(path) => format!("file:/{}", path),
             None => todo!(),
         };
 
+<<<<<<< HEAD
         let subject = S::iri_s2term(&IriS::new_unchecked(&base));
         let graph = Self::load_data_graph(path, &base);
+=======
+        let store = Store::new()?;
+
+        store.bulk_loader().load_graph(
+            BufReader::new(File::open(path)?),
+            RdfFormat::Turtle,
+            GraphNameRef::DefaultGraph,
+            Some(&base),
+        )?;
+
+        let subject = Term::NamedNode(NamedNode::new_unchecked(base.clone()));
+
+        let query = formatdoc! {
+            "
+                SELECT ?this
+                WHERE {{
+                    {} {} ?this
+                }}
+            ", 
+            subject, shacl_validation_vocab::MF_INCLUDE.as_named_node(),
+        };
+>>>>>>> 597911f3ae39befd00a88821b84ec56cdb56faf6
 
         let mut includes = Vec::new();
         for manifest in get_objects_for(