+++ shapes_converter/src/shex_to_html/shex2html_config.rs
@@ -6,12 +6,12 @@ use std::{
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
+#[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
 pub struct ShEx2HtmlConfig {
     pub landing_page_name: String,
     pub css_file_name: Option<String>,
     pub title: String,
-    pub target_folder: PathBuf,
+    pub target_folder: Option<PathBuf>,
     pub color_property_name: Option<String>,
 }
 
@@ -21,7 +21,7 @@ impl Default for ShEx2HtmlConfig {
             title: "Generated shapes".to_string(),
             landing_page_name: "index.html".to_string(),
             css_file_name: Some("shex2html.css".to_string()),
-            target_folder: PathBuf::new(),
+            target_folder: None,
             color_property_name: Some("blue".to_string()),
         }
     }
@@ -29,14 +29,22 @@ impl Default for ShEx2HtmlConfig {
 
 impl ShEx2HtmlConfig {
     pub fn with_target_folder<P: AsRef<Path>>(mut self, target_folder: P) -> Self {
-        self.target_folder = target_folder.as_ref().to_path_buf();
+        self.target_folder = Some(target_folder.as_ref().to_path_buf());
         self
     }
 
+    pub fn target_folder(&self) -> PathBuf {
+        match &self.target_folder {
+            Some(tf) => tf.to_owned(),
+            None => Path::new(".").to_path_buf(),
+        }
+    }
+
     pub fn landing_page(&self) -> PathBuf {
-        self.target_folder
-            .as_path()
-            .join(self.landing_page_name.as_str())
+        match &self.target_folder {
+            Some(tf) => tf.as_path().join(self.landing_page_name.as_str()),
+            None => Path::new(self.landing_page_name.as_str()).to_path_buf(),
+        }
     }
 
     pub fn from_file(file_name: &str) -> Result<ShEx2HtmlConfig, ShEx2HtmlConfigError> {
@@ -46,7 +54,7 @@ impl ShEx2HtmlConfig {
                 error: e,
             }
         })?;
-        serde_yaml::from_str::<ShEx2HtmlConfig>(&config_str).map_err(|e| {
+        serde_yml::from_str::<ShEx2HtmlConfig>(&config_str).map_err(|e| {
             ShEx2HtmlConfigError::YamlError {
                 path_name: file_name.to_string(),
                 error: e,
@@ -63,6 +71,6 @@ pub enum ShEx2HtmlConfigError {
     #[error("Reading YAML from {path_name:?}. Error: {error:?}")]
     YamlError {
         path_name: String,
-        error: serde_yaml::Error,
+        error: serde_yml::Error,
     },
 }