+++ shapes_converter/src/tap_to_shex/tap2shex_error.rs
@@ -1,4 +1,4 @@
-use dctap::{DatatypeId, ShapeId, TapShape};
+use dctap::{DatatypeId, PropertyId, ShapeId, TapShape};
 use prefixmap::PrefixMapError;
 use thiserror::Error;
 
@@ -22,6 +22,9 @@ pub enum Tap2ShExError {
         err: PrefixMapError,
     },
 
+    #[error("No base IRI trying to resolve IRI for {str}")]
+    NoBaseIRI { str: String },
+
     #[error("Multiple value expressions in statement: value_datatype: {value_datatype:?}, value_shape: {value_shape} ")]
     MultipleValueExprInStatement {
         value_datatype: DatatypeId,
@@ -30,6 +33,12 @@ pub enum Tap2ShExError {
 
     #[error("Converting value datatype to IRI, no prefix declaration: {datatype_id:?}")]
     DatatypeId2IriNoPrefix { datatype_id: DatatypeId },
+
+    #[error("Converting shape to IRI, no prefix declaration: {shape_id:?}")]
+    ShapeId2IriNoPrefix { shape_id: ShapeId },
+
+    #[error("Converting property to IRI, no prefix declaration: {property_id:?}")]
+    PropertyId2IriNoPrefix { property_id: PropertyId },
 }
 
 impl Tap2ShExError {