+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## Current changes without release yet
 
+## [0.1.4] - 2024-07-30
+
+- Added options in command line to pass config files in YAML
+- Repaired bug in DCTAP resolution of IRIs
+
+## [0.1.4] - 2024-07-28
+
+- Added 2 separate options for shacl-validate and shex-validate, keeping the generic validate option
+- Repaired bug on UML visualization that didn't show link names
+- Added direct SVG/JPG generation from DCTAP files
+
+## [0.1.3] - 2024-07-27
+
+- Generation of HTML views from ShEx based on Minininja templates which allow better customization
+- Direct conversion from DCTAP to UML and HTML views
+- Generation of UML visualizations in SVG and PNG
+- Basic support for SHACL validation and added shacl-validation crate
+
 ## [0.1.2] - 2024-07-17
 
 - Added descriptions to subcommands in command line